added calculator integration

+++ src/main/java/com/adityabirlacapital/childlifeinsurance/dto/ResponseToAddChildPlanDetails.java
@@ -18,9 +18,9 @@ public class ResponseToAddChildPlanDetails {
 	private Integer childAge;
 	private String goalType;
 	private Double goalAmt;
-	private Integer goalTenure;
-	private Integer tenure;
-	private Integer roi;
+	private Double goalTenure;
+	private Double tenure;
+	private Double roi;
 	private String createdBy;
 	private String modifiedBy;
 	private LocalDateTime modifiedDate;
@@ -30,7 +30,7 @@ public class ResponseToAddChildPlanDetails {
 	 * Need to calculate below fields
 	 */
 	private Integer liChildId;
-	private Double expensesFinal;
-	private Double saveAmount;
+	private Long expensesFinal;
+	private Long saveAmount;
 	private Boolean active;
 }
